Fixed a datatype problem when using GPU (REAL instead of double)

+++ src/viewer/GLViewer.h
@@ -20,9 +20,9 @@ class GLViewer : public QGLWidget, public Viewer
 			);
 
 		void renderFrame (
-				double** U,
-				double** V,
-				double** P,
+                REAL** U,
+                REAL** V,
+                REAL** P,
 				int it
 			);
